requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy[asyncio]
alembic
aiosqlite
python-multipart
python-dotenv
ultralytics
opencv-python
passlib[bcrypt]
python-jose[cryptography]
